+++ CHANGELOG.md
@@ -22,12 +22,13 @@
 - Added classes `StreamCdnCropImageMode` and `StreamCdnResizeImageMode` used for modifying Stream CDN image resize requests. [#4600](https://github.com/GetStream/stream-chat-android/pull/4600)
 
 ### ⚠️ Changed
+- Changed `newMessageIntent` lambda's signature of `NotificationHandlerFactory.createNotificationHandler()`. It receives the whole `Message`/`Channel` entity to help you  create a more complex navigation intent.
 
 ### ❌ Removed
 
 ## stream-chat-android-offline
 ### 🐞 Fixed
-
+- Fixed message gaps in message list view when using offline support. [#4633](https://github.com/GetStream/stream-chat-android/pull/4633)
 ### ⬆️ Improved
 
 ### ✅ Added